pyproject.toml
[project]
name = "scribe-voice"
version = "2.0.0-alpha"
description = "The Open Voice Platform - Your personal memory keeper"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Scribe Contributors", email = "scribe@example.com"}
]
keywords = ["voice", "transcription", "automation", "whisper", "AI", "assistant", "plugins"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Office/Business",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "faster-whisper>=1.0.0",
    "PyQt5>=5.15.0",
    "pynput>=1.7.0",
    "sounddevice>=0.4.0",
    "soundfile>=0.12.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "openai>=1.0.0",
    "psutil>=5.9.0",
    "markdown>=3.0.0",
    "pygetwindow>=0.0.9",
    "ctranslate2>=4.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

media = [
    "pycaw>=20181226",  # Windows audio control
]

[project.urls]
Homepage = "https://github.com/yourusername/scribe"
Documentation = "https://scribe-voice.readthedocs.io"
Repository = "https://github.com/yourusername/scribe"
"Bug Tracker" = "https://github.com/yourusername/scribe/issues"
Changelog = "https://github.com/yourusername/scribe/blob/main/CHANGELOG.md"

[project.scripts]
scribe = "scribe.__main__:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["scribe"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
scribe = ["*.yaml", "*.yml", "assets/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/scribe --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by black)
